# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: InterHive API
  description: InterHive - Industry Readiness & Talent Connect Platform API
  version: 1.0.0
  contact:
    name: InterHive Team
    email: support@interhive.in
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.interhive.in/api/v1
    description: Production Server
  - url: https://staging-api.interhive.in/api/v1
    description: Staging Server
  - url: http://localhost:3000/api/v1
    description: Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        timestamp:
          type: string
          format: date-time
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          minLength: 8
          example: SecurePass123!
    RegisterRequest:
      type: object
      required:
        - email
        - firstName
        - lastName
        - password
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
          minLength: 2
        lastName:
          type: string
          minLength: 2
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [intern, company]
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            accessToken:
              type: string
            refreshToken:
              type: string
            expiresIn:
              type: string

    # User
    User:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, hr, manager, intern, company]
        department:
          type: string
        position:
          type: string
        phone:
          type: string
        profilePhoto:
          type: string
        isActive:
          type: boolean
        isVerified:
          type: boolean
        lastLogin:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Intern
...
  responses:
    UnauthorizedError:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            message: Unauthorized
            timestamp: '2024-01-01T00:00:00.000Z'
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        201:
          description: User registered successfully
        409:
          description: User already exists
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security:
        - bearerAuth: []
      responses:
        200:
          description: Token refreshed successfully
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        200:
          description: User profile retrieved

  # Interns
...
  /attendance/check-in:
    post:
      tags: [Attendance]
      summary: Check in
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                location:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                    address:
                      type: string
                notes:
                  type: string
      responses:
        200:
          description: Check in successful
  /attendance/check-out:
    post:
      tags: [Attendance]
      summary: Check out
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                location:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
                    address:
                      type: string
                notes:
                  type: string
      responses:
        200:
          description: Check out successful
  /attendance/history:
    get:
      tags: [Attendance]
      summary: Get attendance history
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: History retrieved

  # Analytics
...